(Lint) added explicit cast to type to all #definitions

+++ Firmware/Source/ppmdriver.h
@@ -9,7 +9,7 @@
 ///
 /// \file
 ///
-//  Change  #definition of radio channels model moved to config.h
+//  Change: (Lint) added expliicit cast to type to all #definitions 
 //
 //============================================================================*/
 
@@ -24,16 +24,16 @@
 #endif
 #define VAR_GLOBAL
 
-#define PPM_SIGNAL_OK       6   ///< radio signal is OK
-#define PPM_NO_SIGNAL       0   ///< radio signal is bad
+#define PPM_SIGNAL_OK       (uint8_t)6   ///< radio signal is OK
+#define PPM_NO_SIGNAL       (uint8_t)0   ///< radio signal is bad
 
 /* Control mode definitions */
-#define MODE_UNDEFINED      0   ///< mode is undefined
-#define MODE_STAB           1   ///< stabilize roll and pitch
-#define MODE_NAV            2   ///< navigate
-#define MODE_MANUAL         3   ///< manual control
-#define MODE_RTL            4   ///< return to launch
-#define MODE_FPV            5   ///< stabilize camera for FPV
+#define MODE_UNDEFINED      (uint8_t)0   ///< mode is undefined
+#define MODE_STAB           (uint8_t)1   ///< stabilize roll and pitch
+#define MODE_NAV            (uint8_t)2   ///< navigate
+#define MODE_MANUAL         (uint8_t)3   ///< manual control
+#define MODE_RTL            (uint8_t)4   ///< return to launch
+#define MODE_FPV            (uint8_t)5   ///< stabilize camera for FPV
 
 /*----------------------------------- Macros ---------------------------------*/
